+++ packages/upswyng-core/package.json
@@ -20,10 +20,10 @@
     "rrule": "^2.6.2"
   },
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^11.0.0",
     "@types/moment-timezone": "^0.5.12",
     "jest": "^24.9.0",
     "rollup": "^1.27.8",
-    "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-typescript2": "^0.25.3",
