+++ packages/upswyng-server/data-pipeline.rollup.config.js
@@ -1,4 +1,4 @@
-import commonjs from "rollup-plugin-commonjs";
+import commonjs from "@rollup/plugin-commonjs";
 import json from "rollup-plugin-json";
 import pkg from "./package.json";
 import resolve from "rollup-plugin-node-resolve";
