--- packages/upswyng-types/src/TUser.ts
@@ -1,8 +1,0 @@
-export interface TUser {
-  id: string; // database ObjectId converted to hex string
-  name?: string;
-  email: string;
-  providers: ("facebook" | "google")[];
-  isAdmin: boolean;
-  isSuperAdmin: boolean;
-}
